* release-1.23:
  [#1169] Reduce the size of the textarea slightly
  [#1169] Fix misalignment of quick entry metadata section
  [#1169] Fix display of quick buttons/dropdowns section
  [#1169] Tweak placement of the preview/spellcheck buttons
  [#1169] Fix vertical gap when tag wraps onto multiple lines
  [#1169] Widen the max page width for all site skins
  [#1169] Remove gratuitous space between quick-buttons and components
  [#1169] Change the font family of the textarea back to monospace
  [#1169] Shrink font size of entry options (to match rest)
  [#1169] Reduce padding / margins in and around components
  [#1169] Remove space below last form element if it's label + input combo
  [#1169] Tweak the wrapping of the tags checkbox more
  [#1169] Don't force line-height on selects/inputs
  [#1169] Shrink form-element padding inside components
  [#1169] Make sure selects are the same size as the rest of component
  [#1169] Fix wrapping of checkbox/radio text in modals
  [#1169] More wording tweaks
  [#1169] Allow multiple columns in medium-sized screens
  [#1169] Tweak capitalization of "Reading pages"
  [#1169] Change "Crosspost Accounts" to "Crosspost to:"
  [#1169] Fix saving of panel position after moving
  incorporating spacing into the tagnav class instead of having a breakout.
  class handle on the padding, too.
  Additional CSS class "tagnav" to provide customization options for tag arrows
  [#1134] Use protocol-relative URL for image objects in S2
  [#1138] Only add the blank space when we are showing tag arrows
  [#825] Fix URL for renames
  [#1135] Make sure effective remote is set properly for non-BML pages
  [#1002] Don't redirect if we're on $SSLICONPREFIX
  Fix errors found in the logs
  This one too.
  This is not an everywhere table
  Fixes uses of the wrong filter and spurious \

+++ bin/upgrading/update-db-general.pl
@@ -4162,14 +4162,14 @@ EOF
         set_dbnote( "init_media_versions_dimensions", 1 );
     }
 
-    if ( column_type( "renames", "status" ) eq '' ) {
+    if ( table_relevant("renames") && column_type( "renames", "status" ) eq '' ) {
         do_alter( 'renames',
             "ALTER TABLE renames " .
             "ADD COLUMN status CHAR(1) NOT NULL DEFAULT 'A'" );
         do_sql( 'UPDATE renames SET status="U" WHERE renuserid = 0' );
     }
 
-    if ( !check_dbnote( 'remove_countries_from_codes' ) ) {
+    if ( table_relevant("codes") && !check_dbnote( 'remove_countries_from_codes' ) ) {
         do_sql( 'DELETE FROM codes WHERE type = "country"' );
         set_dbnote( 'remove_countries_from_codes', 1 );
     }
